ash: Avoid visual "pop" for launcher scroll gradient fade
If the app launcher has enough apps to scroll we show a gradient at the
bottom of the scroll view. For performance reasons, we defer creation
of the gradient until the launcher show animation is complete. This
creates a subtle visual "pop" at the end of the animation.

On advice from homerr@, create a very quick (50 ms) fade in for the
gradient. This uses LinearAnimation because we can't use layer opacity
animations with a mask layer.

The "pop" is still noticeable if you look for it, but the fade in
makes it look more intentional and less like a bug.

Bug: 1293100
Test: visual
Change-Id: I6fb3b53c1577c182377290818285b7a1c66f4120
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3426436
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966565}