Handle parallel requests to acquire a BluetoothAdapter
Creating and initializing a BluetoothAdapter is an asynchronous
operation and so multiple calls, like bluetooth.getAvailability(),
could be made while it is in progress. This change tightens the
relationship between BluetoothAdapterFactoryWrapper to take advantage of
the guarantee that while AcquireAdapter() may be called more than once
ReleaseAdapter() will not. This makes it okay for these duplicate calls
to be ignored.

Bug: 1233746
Change-Id: I24772d58c0addb72826a375e5a8b44dcb5eb70b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3430518
Reviewed-by: Ovidio Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966581}