Responsive design for ambient animations.
Implements the UX requirements for resizing the animation as a
whole to fit the screen. Luckily, the existing AnimatedImageView
class actually does most of the work for us.

The only thing remaining for responsive design is to add the
portrait-mode version of the animation to the resource pak.
UX has not prepared it yet, but I don't see this as a hish-risk
item.

Performance-wise this had minor impact. We lost 2-3 fps, and I'll
follow up with Skottie experts on this, but 2-3 fps is nowhere near
enough to block this design (that may have even been a transient
result as I've seen the frame rate fluctuate by 2-3 fps in the past).
If anything, we can make up this slight loss easily by refining
other features of the animation that have had much more substantial
impact on performance thus far.

Bug: b:217271296
Change-Id: Ibf97ed8a72a1ca1038cdea2250d7e256b52c3c64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3430901
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#966634}