Streamline file suggestions wrt. suggested content toggle and chips
This is the first of several changes to move files into a more
consistent state between the old and new launcher. Changes:

1. Disable all file suggestion chips in old launcher. These aren't
   launched or under experiment.

2. Never gate ZS files behind the suggested content toggle. This was
   only necessary for chip results.

This does _not_ change the querying behavior for Drive. It's important
we keep this constant so we have a fair QPS comparison between M99 and
M100.

Bug: 1291866
Change-Id: I240a022c07577f8ff9319916024fa9a9973ba528
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3435245
Reviewed-by: Rachel Wong <wrong@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966645}