Rearranges some methods in AXPosition and renames others for clarity
Our goal is to expose the same functionality that is found in AXPosition,
(which is used by internal accessibility logic), to AutomationPosition
which is publicly available as an API on Chrome OS.
To ensure that "no method is left behind" this patch rearranges the methods
in AXPosition   to be in the same order as they would ultimately be in the
public API.
It also takes the opportunity to rename AsTextSelectionPosition to ASDomSelectionPosition
and LowestCommonAncestor to LowestCommonAncestorPosition,
which better describes the methods' purpose and makes
their naming more consistent throughout the class.
A ToString  method
was added to allow converting AXPositionKind
to a human-readable string. Again, this functionality will be used in the publicly
available AutomationPosition class.

AX-Relnotes: n/a.
Change-Id: I319b8e91db36a6addfa1ea1568bcabda07f561a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3427507
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966647}