Move impl of ReferenceOutput::Listener::OnPlayoutData to separate class
audio::OutputTapper currently manages both subscription to and reception of the audio output reference.
This CL moves the ReferenceOutput::Listener implementation to a separate class, AudioProcessorHandler.

This is part of a larger effort to move input audio processing effects (noise suppression, echo cancellation, etc) into the audio process.
The AudioProcessorHandler will manage the audio processing effects for the InputController.

The OutputTapper may be used even when no AudioProcessorHandler is present. This is handled by adding a no-op implementation of ReferenceOutput::Listener.

A note on names: OutputTapper was up until recently named AudioProcessorHandler, but has been renamed to better reflect the setup in this CL.
https://chromium-review.googlesource.com/c/chromium/src/+/3420493

Bug: chromium:1215061
Change-Id: Iff5cf14dd79ae4227cca92dfa5666374f9ecddb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3420412
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Sam Zackrisson <saza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966655}