DanglingPtr: Fix WatcherDispatcher
This is the #2 source of dangling raw_ptr: 31% of the occurrences,
when running ContentShell.

This is not meant to be dereferenced. This is a dangling pointer at its
core. It shouldn't hold back memory though, it is used only for
comparisons.

Fixing this avoid going through the path of putting the memory region
under quarantine and the cost of memsetting the memory on free.

This patch replaces it by a uintptr_t, this fix the issue, because
contrary to raw_ptr it doesn't increment the refcount. Moreover, this
makes it clear it doesn't have to be dereferenced.

See doc:
https://docs.google.com/document/d/15qq2wr7Xza5awgJAfXUzO1UZFAG7lXDm4K8hd7vC3GU/edit#heading=h.p5mua6e7y11x

Dangling ptr preliminary list:
https://docs.google.com/spreadsheets/d/1BAfz1uHhMkFe2sK2vyTHsJ733VQUhY9v2M4HxOwaM7E/edit?resourcekey=0-H9SwkJTMbBxQPKUEuIrwow#gid=0

See also:
https://docs.google.com/document/d/1fZN-7def91IgXUexhg-8bu2F6Lo7YrXcy0v3rbjZRNg/edit

Bug: 1291138
Change-Id: I113a67fe2f88369cae883d50d2a968c60f890aa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3424045
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966715}