Moving the ProtocolHandlerThrottle to //components
This CL is part of the ongoing refactoring we are doing to move the
protocol handler logic to a new Custom Handler //component.

Inn order to move the ProtocolHandlerRegistry's browser tests we need
the ShellContentBrowserClient to implement the CreateURLLoaderThrottle
public API, in a similar way that the ChromeContentBrowserClient does.

This is a prior step to this abstract public API implementation, so
that both ContentBrowserClient implementations can share the throttle
logic to deal with custom protocol handlers.

This CL doesn't implement any behavior change, so I consider that no
additional tests are needed.

Bug: 1292925, 1139176
Change-Id: I230bc5a31b3b992754473d263028cee43162fe64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3427644
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#966731}