[Nearby] Check background scanning flag before starting cooldown timer
Previously, we were starting the fast initiation cooldown timer after a
successful transfer even if the background scanning feature flag was
disabled. This resulted in users being shown the background scanning
notification even when the feature was disabled. If the user clicked on
the notification, they would de-reference a null pointer and crash.

In this CL, we check that the background scanning feature is enabled
before starting the timer. As a failsafe, we add another feature state
check to InvalidateFastInitiationScanning(), and we early return if
the feature is disabled.

The cooldown timer addition landed in 99.0.4808.0.
(See https://crrev.com/c/3353887.)

Bug: b/216160852
Change-Id: I5508621b65d157d573552d4294112656ebfec7fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3434568
Reviewed-by: Claude van der Merwe <cvandermerwe@google.com>
Auto-Submit: Josh Nohle <nohle@chromium.org>
Reviewed-by: Josh Nohle <nohle@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966739}