ozone/wayland: fix gpu hang when switching to sw compositing
When GPU fails and the service switches from hw accelerated
compositing to software one, users may experience frozen
graphics with just a small fraction of the browser window
updated.

The analysis has shown that there is a subsurface that remains
on top of the root surface (that is, WaylandCanvasSurface attaches
buffers to the root surface that plays role of a background surface
for the hw accelerated compositing case), and it has the previous
buffer attached. That results in a Wayland compositor (exo, or mutter,
maybe others), keeping the reference to the buffer and showing that.

Debugging revealed two problems - one with the WaylandFrameManager that
did not attach a NULL buffer to a subsurface if a frame did not have
submitted buffers stored as they can either be erased during
WaylandFrameManager::ClearState or if a Wayland compositor releases
them.
And the second problem was caused by WaylandWindow::ClearState using
only non-primary subsurfaces for an empty frame.

Test: WaylandBufferManagerTest.AttachesNullBuffersOnChannelDestroyed
Bug: 1201314
Change-Id: If30f7a50f08127c1f83fc414d9cf2acfbaa3a170
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3423831
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#966743}