[iOS] ContextMenu image preview animation bugfix
Context menu animation was glitching for images, it was caused by
setting the view preferred size during the animation.
The image is loaded asynchronously
and we only knew its size after loading it:
- Set view default size
- Start image loading
- Start view animation
- End image loading
- Set view preferred size // Glitch here
- End view animation
Now we get the image size with a javascript callback before the
animation. So the view's preferred size doesn't change during the
animation.

Bug: 1288017
Change-Id: Ic0db9f0278ea22b053f8f96b2612712ffe82037e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3432927
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966746}