Don't put a null string into ManifestDebugInfo::raw_manifest.
This is a speculative fix for a invalid message crash bug in
ManifestDebugInfo. There is a case where the manifest fetch can
succeed but have a null data string. This string was being set
on ManifestDebugInfo, which would result in a failure to
validate due to the raw_manifest field expecting a non-null
input.

Bug: 1291029
Change-Id: I511683ed858823649c6642397311731b1a50a1eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428390
Auto-Submit: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966765}