Test that net::AddressList::Deduplicate() preserves original order.
I tried to be clever by simplifying the definition of Deduplicate(), but
my change would have broken the order-preservation promise.

The silver lining is that this revealed a deficiency in our test
coverage. The existing tests did not fail for my "improved"
Deduplicate() because they only try inputs where the subsequence of
unique elements is already sorted.

This CL now adds a test that tries all permutations of a small list of
IPEndPoints.

Change-Id: Ic64bd3052169218a3f3f84d2468a30a7cd7be35e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433206
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966803}