[FLEDGE] Minor AuctionRunner cleanups.
Remove AuctionWorkletManager::Delegate and frame origin from
AuctionRunner, as it no longer needs them. They're both now implicitly
bound to the AuctionWorkletManager instead, which is a creation
parameter to the AuctionRunner.

Add option for the buyers vector in AuctionRunner tests, to be nullopt,
since that is case is now handled by the AuctionRunner, and add a unit
test for it (There's already an integration test, but should have a
simpler test for that case, and plan to add a component auction unit
test only for the corresponding component auction case).

Bug: 1288865
Change-Id: I588f8da86dd34c0718cccc634cdd86454a1ef955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3430683
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966807}