Refactor WebContentsAccessibilityTest and add more accessibility tests
-------------------------------------------------------

To sheriffs: If a test fails/flakes, consider adding the "@DisabledTest"
             annotation to the test in lieu of a full revert.
             Thanks!

Test file location:
...content/browser/accessibility/WebContentsAccessibilityTest.java

-------------------------------------------------------

This CL follows-up the 'Missing Layer' work with a refactor of the
existing tests of WebContentsAccessibilityTest.

With this CL we clean up the previously existing tests to make them
easier to understand and in-line with newer standards we added during
the missing layer work. We also add two new tests for better coverage:
   testResetFocus
   testPerformAction_showOnScreen

AX-Relnotes: N/A
Bug: 1258230
Change-Id: Iac9ef4f4428ddb732664582c79b03a8b10d8127b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428484
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#966876}