Align base::win::GenericScopedHandle APIs with base::ScopedGeneric's.
base::ScopedPlatformFile is implemented with
base::win::GenericScopedHandle and base::ScopedGeneric, on Windows and
POSIX, respectively. When a cross-platform caller gets back a
base::ScopedPlatformFile, the caller cannot easily use the
base::ScopedPlatformFile in some cases because the APIs for the
platform-specific implementions are named differently.

Improve this situation with the following renames to line up the APIs:
IsValid() -> is_valid()
Get()     -> get()
Take()    -> release()

Keep methods with the old names as wrappers around the actual
implementation, to avoid breaking the build. They will be removed
separately.

Bug: 1291793
Change-Id: I9b41edc9007a126a315ff657578f26f3875bce5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3426390
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966896}