mac: Package install_name_tool in the hermetic Xcode package.
An upcoming project (see bug) will need to link the same .dylib
to images at different relative path depths, which means the
recorded install name of each linked image will need to be
adjusted for the correct depth. It is not possible to specify
an alternate install name at link time, so the install_name_tool
is used to do that.

Bug: 1255223
Change-Id: Ia378a5a88a0fcdd59f26f85906501f44811bb25c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428498
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966916}