[Extension bindings] Add success/failure callbacks to HandleRequest
This CL changes the way we call JS HandleRequest custom hooks to pass a
separate success and failure callback. This lets us handle logic for
rejecting the associated promise or setting lastError and calling the
callback outside of the custom hook itself.

Part of this also involved refactoring HandleRequest custom hooks with
asynchronous responses to actually use the already existing logic in
APIRequestHandler and remove the similar logic that was being done to
resolve requests in APIBindingHooks previously.

To implement this, logic also had to be added to APISignature to handle
the handful of APIs which have trailing function parameters that are not
actually meant to be treated as success callbacks. Once all API schemas
have been moved over to use the returns_async format, this logic will be
able to be removed as it will be explicit when an API has an asynchronous
return and not need to be inferred from the trailing function parameter.

Bug: 1220910
Change-Id: I72c6825374fd501e831f4c8a4ab2be54a7be5689
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3233175
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966938}