Dictation: Allow Dictation UI to be shown when ChromeVox is active
This change allows the Dictation UI to be shown when ChromeVox is
active. Previously, we disallowed this behavior because we were using
the IME window as a temporary UI, which caused a lot of noisiness with
ChromeVox. This change also ensures that ChromeVox announces Dictation
UI hints when they appear by sending an alert event to ChromeVox.

AX-Relnotes: N/A
Bug: 1252037
Change-Id: I7a4f327908da4b701a04769ca1d3f2875120e41e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3434288
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966954}