Remove outdated TODO in SiteInstanceImpl
SiteInstances still support an initial state with no SiteInfo, and will
likely continue to do so. This allows creating renderer processes and
blank RenderFrames that are ready to go for a future navigation in most
cases (in scenarios like the Android NTP). The alternative would be
delaying having a SiteInstance until the SiteInfo is known, but that
would be more problematic for RenderFrameHost and for process creation.

We now have sufficient safeguards to prevent using an unsuitable
process, and the need to discard this blank process should be rare.

Bug: 1015882
Change-Id: I56f2c881fcd4f6ef1e02e7c176488b028cc4c7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433275
Reviewed-by: Charles Reis <creis@chromium.org>
Commit-Queue: Sharon Yang <yangsharon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966963}