Prevent tautological comparisons to zero in unsigned types
We want to avoid code that was of the form:
  for (int i = size; i >= 0; --i)

From being converted to size_t and becoming an infinite loop without
any warning from the compiler.

However Wtautological-unsigned-zero-compare is:
 - Not available on NaCl or Windows compilers.
 - Produces errors due to chromium's use of unqualified
   chars (which are sometimes signed) being compared to
   to 0 (which is correct when char is signed).

Thus, for now we scope the warning to the prevent_unsafe_narrowing
config, but we need to either build with -funsigned-char or
remove chars from the Wtautological-unsigned-zero-compare warning
and enable it globally.

Bug: 1292951
Change-Id: I8bc7901c48fe00e834b0bc8df5081d2b53e50fa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3429460
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966988}