Dictation: Fix standby icon alignment issues
This change applies a transform to the Dictation standby icon so that
it fits into a 22 x 22 box. The previous icon was drawn in a 22 x 14
box, which resulted in alignment issues when rendering the icon in
an ImageView (which expects an icon to be square). Please see the before
and after screenshots of the icon.

Before:
https://drive.google.com/file/d/1mEbQgiFQui685691Zeu4JmZoFCKrhhfQ/view?usp=sharing&resourcekey=0-GGzi4Aeccpbl6urkzxEJVg

After:
https://drive.google.com/file/d/1xuwffmfiMQOUa2wvklzLGLUUrgia1ZXh/view?usp=sharing&resourcekey=0-g9mAQcEobjcPKZU1ZNUYbg

AX-Relnotes: N/A
Bug: 1252037
Change-Id: I648673cc261ce97fe47707baa2294fee5303284c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438128
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966997}