Fix activation state consistency in DesktopNativeWidgetAura
This CL introduces introduces a boolean flag `should_activate_` to
DesktopNativeWidgetAura which is used to determine whether the widget
should receive activation in DNWA::ShouldActivate().

This flag is set when handling an activation change event in
DNWA::HandleActivationChanged().

The flag is used to ensure that the activation client correctly
sees whether the widget and any associated windows should receive
activation when handling the activation change event.

This is necessary as the widget's delegate may not have been notified
of the activation state change before the activation client updates
its state.

This CL also addresses a bug where activation changes in the desktop
native widget was not being propagated to delegates under certain
circumstances.

Bug: 1284537
Change-Id: Ie70bcf9534ba18731e03f8a1aff14e32d0c0a4c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3426083
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967009}