Implement disconnectable client
This is a wrapper on top of an abstract client which may become
available or unavailable at any moment. Calls made through the client
are rejected with error if the client is not available, and cancelled
immediately returning an error, if it becomes unavailable while the call
is in progress.
The wrapper is intended to be used for dBus connection to missive client
in Chrome and from daemon to Chrome in missived.

Bug: b:205814822
Change-Id: Ieeff24e70dbf7fbc50ae91012517ccb4ca53a7f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3437863
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967041}