multiword: reuse IsPredictiveWritingEnabled
Previously in crrev/c/3413633 a default value was added for the
predictive writing pref when the user has not interacted with the
relevant toggle in language settings.

The CL however, did not cover the both instances where we inspect the
predictive writing pref.

This CL in light of this ensures we use the same method in both
instances, instead of having two seperate functions implementing the
same behaviour.

BUG=b:216047778

Change-Id: I730d6e602b9526bf301e8ad481006b3b76385fe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3434687
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967061}