DanglingPtr: Fix sql StatementRef::Close()
In the latest check on:
chrome/browser/chrome_web_platform_security_metrics_browsertest.cc
this seems to be the #1 source of dangling raw_ptr: 21.57%.

Nothing wrong with that, because the data is not accessed later.
However, this forces holding back memory under quarantine and the cost
of memsetting the memory on free. Given the high number of occurences,
it is worth tweaking.

The high usage might be explained by the nature of the test or might be
representative of the general behavior. I am not sure.

See doc:
https://docs.google.com/document/d/15qq2wr7Xza5awgJAfXUzO1UZFAG7lXDm4K8hd7vC3GU/edit#heading=h.p5mua6e7y11x

Dangling ptr preliminary list:
https://docs.google.com/spreadsheets/d/1BAfz1uHhMkFe2sK2vyTHsJ733VQUhY9v2M4HxOwaM7E/edit?resourcekey=0-H9SwkJTMbBxQPKUEuIrwow#gid=767599093

See also:
https://docs.google.com/document/d/1fZN-7def91IgXUexhg-8bu2F6Lo7YrXcy0v3rbjZRNg/edit

Bug: 1291138
Change-Id: I3237cbe793d64a2a5533456e8d527ec9ed84f9f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3424056
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967084}